Extract the parsing/construction from LockfileEntry.ts and LockfileDependency.ts into lfxGraphLoader.ts (preparing to later move this code from client to server)
